refactor based on go-standard & use ctx for cli
