--- cmd/cmd.go
@@ -1,6 +1,0 @@
-package cmd
-
-func Execute() error {
-	err := rootCmd.Execute()
-	return err
-}
